Move FloatSupport into adapter namespace
